resignation.js -> employee.js로 통합, 사원 세부 정보 detail로 통합
