+++ src/main/resources/templates/employee/resignation-form.html
@@ -14,66 +14,58 @@
             <table>
                 <tr>
                     <th>사원 ID</th>
-                    <td th:text="${employee.employeeId}">사원 ID</td>
+                    <td th:text="${employee.employeeId}"></td>
                 </tr>
                 <tr>
                     <th>이름</th>
-                    <td th:text="${employee.name}">이름</td>
+                    <td th:text="${employee.name}"></td>
                 </tr>
                 <tr>
                     <th>주민등록번호</th>
-                    <td th:text="${employee.birth} + '-' + ${employee.residentRegistrationNumber}">주민등록번호</td>
+                    <td th:text="${employee.birth} + '-' + ${employee.residentRegistrationNumber}"></td>
                 </tr>
                 <tr>
                     <th>전화번호</th>
-                    <td th:text="${employee.phone}">전화번호</td>
+                    <td th:text="${employee.phone}"></td>
                 </tr>
                 <tr>
                     <th>이메일</th>
-                    <td th:text="${employee.email}">이메일</td>
+                    <td th:text="${employee.email}"></td>
                 </tr>
                 <tr>
                     <th>주소</th>
-                    <td th:text="${employee.address}">주소</td>
+                    <td th:text="${employee.address}"></td>
                 </tr>
                 <tr>
                     <th>상세 주소</th>
-                    <td th:text="${employee.detailAddress}">상세 주소</td>
+                    <td th:text="${employee.detailAddress}"></td>
                 </tr>
                 <tr>
                     <th>부서</th>
-                    <td th:text="${employee.department.name}">부서</td>
+                    <td th:text="${employee.department.name}"></td>
                 </tr>
                 <tr>
                     <th>직급</th>
-                    <td th:text="${employee.position.name}">직급</td>
+                    <td th:text="${employee.position.name}"></td>
                 </tr>
                 <tr>
-                    <th>입사일</th>
-                    <td th:text="${employee.hireDate}">입사일</td>
+                    <th>입사 날짜</th>
+                    <td th:text="${employee.hireDate}"></td>
                 </tr>
                 <tr>
-                    <th>상태</th>
-                    <td th:text="${employee.status}">상태</td>
+                    <th>재직 상태</th>
+                    <td th:text="${employee.status}"></td>
                 </tr>
                 <tr>
                     <th>남은 연차</th>
-                    <td th:text="${employee.remainingLeave}">남은 연차</td>
-                </tr>
-                <tr>
-                    <th>마지막 수정일</th>
-                    <td th:text="${employee.lastModified}">마지막 수정일</td>
-                </tr>
-                <tr>
-                    <th>수정자</th>
-                    <td th:text="${employee.modifiedBy}">수정자</td>
+                    <td th:text="${employee.remainingLeave}"></td>
                 </tr>
             </table>
-            </div>
+        </div>
 
         <div>
             <!-- 퇴사 처리 폼 -->
-            <form th:action="@{/employee/resign/{employeeId}(employeeId=${employee.employeeId})}" method="post" onsubmit="submitResignationForm(event)" enctype="multipart/form-data">
+            <form th:action="@{/employee/resign/{employeeId}(employeeId=${employee.employeeId})}" onsubmit="submitResignationForm(event)" enctype="multipart/form-data">
                 <div class="file-area">
                     <label for="resignationDocuments">첨부 파일 (최대 3개, 각 10MB 이하)</label>
                     <input type="file" id="resignationDocuments" name="resignationDocuments" onchange="addFile(this)" multiple/>
