+++ src/main/resources/templates/employee/resignation.html
@@ -6,7 +6,12 @@
 <th:block layout:fragment="css"></th:block>
 
 <th:block layout:fragment="script">
-    <script th:src="@{/js/resignation.js}"></script>
+    <script th:src="@{/js/employee.js}"></script>
+    <script>
+        document.addEventListener("DOMContentLoaded", function() {
+
+        });
+    </script>
 </th:block>
 
 <div layout:fragment="~{content}">
@@ -71,7 +76,7 @@
                 <td th:text="${employee.resignation?.codeNumber}">퇴사 코드</td>
                 <td th:text="${employee.resignation?.specificReason}">구체적 사유</td>
                 <td>
-                    <button type="button" th:employeeId="${employee.employeeId}" th:onclick="openModal('/employee/resignation-detail/' + this.getAttribute('employeeId'))">상세 정보 조회</button>
+                    <button type="button" th:employeeId="${employee.employeeId}" th:onclick="openModal('/employee/' + this.getAttribute('employeeId'))">상세 정보 조회</button>
                 </td>
             </tr>
             </tbody>
@@ -95,7 +100,7 @@
             </tr>
             </thead>
             <tbody>
-            <tr th:each="employee : ${preDeletionEmployees}">
+            <tr th:each="employee : ${preDeletionEmployees}" th:employeeId="${employee.employeeId}" th:onclick="openModal('/employee/' + this.getAttribute('employeeId'))" style="cursor: pointer;">
                 <td th:text="${employee.employeeId}">사원 ID</td>
                 <td th:text="${employee.name}">이름</td>
                 <td th:text="${employee.department}">부서</td>
@@ -107,7 +112,7 @@
                 <td th:text="${employee.resignation?.codeNumber}">퇴사 코드</td>
                 <td th:text="${employee.resignation?.specificReason}">구체적 사유</td>
                 <td>
-                    <button type="button" class="delete-btn" th:employeeId="${employee.employeeId}">삭제하기</button>
+                    <button type="button" class="delete-btn" th:employeeId="${employee.employeeId}" th:employeeName="${employee.name}" onclick="submitDeleteRequest(event)">삭제하기</button>
                 </td>
             </tr>
             </tbody>
