+++ src/main/resources/templates/fragments/layout-with-modal.html
@@ -60,6 +60,8 @@
                 } else {
                     console.error('First input field not found');
                 }
+
+                updateDateTimeElements(); // DateTime T 제거
             });
     }
